Readme updated

+++ routes/api/product-routes.js
@@ -59,16 +59,10 @@ router.get('/:id', (req, res) => {
 
 // create new product
 router.post('/', (req, res) => {
-  Product.create({
-    product_name: req.body.product_name,
-    price: req.body.price,
-    stock: req.body.stock,
-    category_id: req.body.category_id,
-    tagIds: req.body.tagIds
-  })
+  Product.create(req.body)
   .then((product) => {
     // if there's product tags, we need to create pairings to bulk create in the ProductTag model
-    if (req.body.tagIds.length) {
+    if (req.body.tagIds && req.body.tagIds.length) {
       const productTagIdArr = req.body.tagIds.map((tag_id) => {
         return {
           product_id: product.id,
@@ -95,38 +89,34 @@ router.put('/:id', (req, res) => {
       id: req.params.id,
     },
   })
-    .then((product) => {
-      // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
-      // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // create filtered list of new tag_ids
+  .then((product) => {
+    if (req.body.tagIds && req.body.tagIds.length) {
+      const productTags = ProductTag.findAll({ where: { product_id: req.params.id }})
+
+      const productTagIds = productTags.map(({ tag_id }) => tag_id)
+
       const newProductTags = req.body.tagIds
         .filter((tag_id) => !productTagIds.includes(tag_id))
         .map((tag_id) => {
           return {
             product_id: req.params.id,
             tag_id,
-          };
-        });
-      // figure out which ones to remove
+          }
+        })
       const productTagsToRemove = productTags
         .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
+        .map(({ id }) => id)
 
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
-    })
-    .then((updatedProductTags) => res.json(updatedProductTags))
-    .catch((err) => {
-      // console.log(err);
-      res.status(400).json(err);
-    });
+        return Promise.all([
+          ProductTag.destroy({ where: { id: productTagsToRemove }}),
+          ProductTag.bulkCreate(newProductTags)
+        ])
+    }
+    return res.json(product)
+  })
+  .catch((error) => {
+    res.status(400).json(error)
+  })
 });
 
 router.delete('/:id', (req, res) => {
